+++ src/interface/IFileExpirations.ts
@@ -1,0 +1,8 @@
+import IFileType from "./IFileType";
+
+interface IFilesExpiration {
+	default?: string;
+	fileTypes?: IFileType;
+}
+
+export default IFilesExpiration;