+++ src/interface/IFileType.ts
@@ -1,0 +1,5 @@
+interface IFileType {
+	[key: string]: string;
+}
+
+export default IFileType;