+++ src/interface/IFilesExpiration.ts
@@ -1,5 +1,0 @@
-interface IFilesExpiration {
-	[key: string]: string;
-}
-
-export default IFilesExpiration;