+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New keywords will be quoted for you in the generated `Content-Security-Policy` header value of the `contentSecurityPolicy` option: - `unsafe-eval`, - `unsafe-hashes`, - `unsafe-inline`, - `strict-dynamic`, - `report-sample`.
+- New option, `fileExpirations`, to let you control how you file types should be cached by the browser (see the documentation for the description of the API).
 
 ### Added